:sparkles: Always encode signatures
This commit packs a good amount of changes, yet the most important one
is that now we always encode the salt and signature and this allows
the high-level class to verify the signature before decoding (this is
not _entirely_ true, since it receives string and it needs bytes so it
needs to `.encode()` it before checking the sig, but there's no way
around this unfortunately: it's the simplest operation that can be
achieved before checking sig; i.e. other packages even decodes base64
data apart from converting string to bytes).

The salt is actually generated and used as base64 data to avoid
needing to decode it when checking the signature, which would
otherwise defeat this whole refactor. This causes the loss of a few bits
of random data but its otherwise negligible given the largest possible
salt size is used.

To separate the composite signature (salt + signature) from the data a
symbol (`.`) is used given that now it's not possible to know the fixed
size from beforehand.

The timestamp value is also base64 encoded and decoded, but that's
relatively safe since the signature is verified before decoding.
Similarly as before a symbol (`.`) is used to separate the timestamp
from the data.

Force bytes was removed from low-level functions because I can't help
myself to see that there.

Blake2Serializer was renamed to Blake2SerializerSigner because of
reasons.