:recycle: Refactor all exceptions
Split exception into something more useful so now it's easier to catch
exactly what you want: exceptions are diveded between those that are
related to working with signed data, those that are related to unsigned
data and those that not related to data at all.

Improve descriptions in docstrings and make all exceptions properly
explicit.