:building_construction: Split serializer class into mixins
Instead of inyecting objects to the serializer class, inherit from the
timestamp abstract class simplifying the code.

Additionally split into mixins and abstracts so that with this
construction we allow end users the possibility to build and adapt new
classes for their own use case, as well as to easily implement custom
serializer such as msgpack or custom compressor and encoder, i.e.
base62.

Allow to set the desired compression level.

Check and skip detrimental compression.