:sparkles: Add support for BLAKE3
BLAKE3 is available through the third-party package `blake3`. This
commit adds support for it, ensuring everything works without it, too.

Additionally, add hasher variation in tests to ensure that everything
works, in case the code paths become different in the future (they are
still mostly the same for every hasher).