:memo: Add deterministic sigs in comparison docs
Deterministic signatures are faster, in the sense that there's no need
to generate a salt. This commit adds signers in deterministic mode in
the packages comparison, which is actually the correct way to compare
given that the others don't use a random salt by default.

Additionally, add a link to the performance comparison of deterministic
vs non-deterministic in the details doc.