:zap: Replace frozen dataclass for NamedTuple
This commit is a micro-performance improvement. It changes the internal
and external data containers from frozen dataclass to typing.NamedTuple
that, for our use case, are equivalent. This is done for performance
reasons: instantiating a frozen dataclass takes as much as two times as
NamedTuple. And we lose nothing in exchange; we even gain a new possible
usage: `sig, data = signer.sign_parts(data)`.