:label: Replace AnyStr for signer's methods
This commit changes the usage of AnyStr in signer's public methods with
Union[str, bytes]. It is not the same, and they are not generally
interchangeably like this, but for this particular usage Union makes
more sense, and MyPy seems to agree with this.
Using AnyStr was probably wrong in the first place, although it mostly
did not, and does not make any difference, which is why this will be
part of the next minor release.