:hammer: Set signing timestamp on tasks
This commit will set the timestamp for the create-tag and
create-archives tasks first and foremost, to the tag creation value if
possible, or otherwise to the current time.
This ensures that all signatures will have the same timestamp.

Additionally, now all archives and its signatures are created in a
subdirectory instead of root.